PythonExamples/web_api/python-flask-server-generated/swagger_server/controllers/default_controller.py
import connexion
import six

from swagger_server import util


def histogram_get():  # noqa: E501
    """Get Histogram

    Retrieve a histogram of passenger data. # noqa: E501


    :rtype: None
    """
    return 'do some magic!'


def passengers_get():  # noqa: E501
    """Get All Passengers

    Retrieve a list of all passengers. # noqa: E501


    :rtype: None
    """
    return 'do some magic!'


def passengers_id_get(id, attributes):  # noqa: E501
    """Get Passenger Data by ID

    Retrieve passenger data by ID with specific attributes. # noqa: E501

    :param id: ID of the passenger to retrieve.
    :type id: int
    :param attributes: Comma-separated list of attributes to include in the response.
    :type attributes: List[str]

    :rtype: None
    """
    return 'do some magic!'
